Isolates maven integration dependant calls
